requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning and Transformers
torch>=2.6.0
torchvision>=0.21.0
transformers>=4.37.0
accelerate==0.24.0
sentencepiece==0.2.0
protobuf==4.25.0
tiktoken>=0.5.0

# RAG and Vector Search
sentence-transformers==2.2.2
faiss-cpu==1.7.4  # Use faiss-gpu if GPU is available
numpy==1.24.3

# HTTP Requests and Testing
requests==2.31.0
colorama==0.4.6

# Logging and Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Optional: For better performance
# bitsandbytes==0.41.2  # For 8-bit quantization
# optimum==1.14.0  # For model optimization